requirements.txt
# Web Framework
fastapi>=0.109.2        # Latest stable with improved middleware support
uvicorn[standard]>=0.27.1  # Latest with improved logging
python-multipart>=0.0.7    # For file uploads
python-jose[cryptography]>=3.3.0  # For JWT
passlib[bcrypt]>=1.7.4     # For password hashing

# Database
sqlalchemy>=2.0.27         # Latest with improved async support
alembic>=1.13.1           # Latest database migration tool

# AI and Document Processing
llama-index-core>=0.12.10      # Core LlamaIndex functionality
llama-index-readers-file>=0.1.0 # For file reading capabilities
llama-index-embeddings-openai>=0.1.0  # For OpenAI embeddings
llama-index-vector-stores-qdrant>=0.1.0  # For Qdrant vector store
llama-index-llms-openai>=0.1.0  # For OpenAI LLM integration
openai>=1.12.0            # Latest stable with streaming improvements
pypdf>=3.17.4             # For PDF processing

# Vector Store
qdrant-client>=1.7.0      # Latest with improved batch operations

# Core Dependencies
pydantic>=2.6.1           # Latest with improved validation
python-dotenv>=1.0.0      # For environment variables
typing-extensions>=4.9.0   # For enhanced type hints
python-dateutil>=2.8.2    # For date parsing
requests>=2.31.0          # For HTTP requests
